# backend/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: MERMAID API
  version: 1.0.0
  description: |
    MERMAID (Marine Safety and Market Information Dashboard) API.
    MVP scope includes authentication, advisories, marine conditions,
    vendor demand listings, fisherman marketplace, trip sessions, and catch logs.
servers:
  - url: http://localhost:8080/api
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /buyer/favorites:
    get:
      tags: [Buyer Favorites]
      summary: List buyer's favorites (filterable by target type)
      operationId: getBuyerFavorites
      parameters:
        - in: query
          name: type
          schema:
            $ref: '#/components/schemas/FavoriteTargetType'
      responses:
        '200':
          description: Favorites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuyerFavorite'
    post:
      tags: [Buyer Favorites]
      summary: Save a listing or vendor as favorite (idempotent)
      operationId: addBuyerFavorite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFavoriteRequest'
      responses:
        '201':
          description: Favorite created or already existed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerFavorite'
        '400':
          description: Invalid target
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /buyer/favorites/{favoriteId}:
    delete:
      tags: [Buyer Favorites]
      summary: Remove a favorite by id
      operationId: removeBuyerFavorite
      parameters:
        - in: path
          name: favoriteId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Removed
  /buyer/favorites/by-target:
    delete:
      tags: [Buyer Favorites]
      summary: Remove a favorite by target type + target id (convenience for unheart on cards)
      operationId: removeBuyerFavoriteByTarget
      parameters:
        - in: query
          name: targetType
          required: true
          schema:
            $ref: '#/components/schemas/FavoriteTargetType'
        - in: query
          name: targetId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Removed (no-op if not present)
...
  /buyer/orders:
    post:
      tags: [Buyer Orders]
      summary: Place an order against a vendor demand listing
      operationId: placeBuyerOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyerPlaceOrderRequest'
      responses:
        '201':
          description: Order placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Listing not found or closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Buyer Orders]
      summary: List buyer's own orders
      operationId: getBuyerOrders
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED, DISPUTED]
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /buyer/orders/{orderId}:
    get:
      tags: [Buyer Orders]
      summary: Get a single buyer order by ID
      operationId: getBuyerOrderById
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found or not owned by this buyer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /buyer/orders/{orderId}/timeline:
    get:
      tags: [Buyer Orders]
      summary: Get order status timeline (Buyer)
      operationId: getBuyerOrderTimeline
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Timeline events for the order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderStatusEvent'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /uploads:
    post:
      tags: [Uploads]
      summary: Upload a file (image)
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                subDir:
                  type: string
                  description: Subdirectory (avatars, listings, reviews)
                  default: general
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid file type or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /notifications:
    get:
      tags: [Notifications]
      summary: Get notifications for current user
      operationId: getNotifications
      parameters:
        - in: query
          name: unreadOnly
          schema:
            type: boolean
            default: false
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
  /notifications/{notificationId}/read:
    put:
      tags: [Notifications]
      summary: Mark a notification as read
      operationId: markNotificationRead
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Notification marked as read
...
  /buyer/profile:
    get:
      tags: [Buyer Profile]
      summary: Get buyer profile
      operationId: getBuyerProfile
      responses:
        '200':
          description: Buyer profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerProfile'
    patch:
      tags: [Buyer Profile]
      summary: Update buyer profile
      operationId: updateBuyerProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyerProfileUpdateRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerProfile'
  /buyer/orders/{orderId}/payment-intent:
    post:
      tags: [Buyer Orders]
      summary: Create a payment intent for an order
      operationId: createPaymentIntent
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Payment already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Role:
      type: string
      enum: [ADMIN, VENDOR, FISHERMAN, BUYER]
...
    DemandListingStatus:
      type: string
      enum: [OPEN, CLOSED]
...
    FishSpecies:
      type: object
      required: [id, commonName, active]
      properties:
        id:
          type: integer
          format: int64
        commonName:
          type: string
        scientificName:
          type: string
          nullable: true
        active:
          type: boolean
...
    MarketLocation:
      type: object
      required: [id, name, municipality, active]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        municipality:
          type: string
        province:
          type: string
          nullable: true
        active:
          type: boolean
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
    BuyerPlaceOrderRequest:
      type: object
      required: [listingId, dispatchMode]
      properties:
        listingId:
          type: integer
          format: int64
        orderedQtyKg:
          type: number
          format: double
          nullable: true
        orderedQtyEstimate:
          type: string
          nullable: true
        dispatchMode:
          type: string
          enum: [PICKUP, DELIVERY]
        deliveryAddress:
          type: string
          nullable: true
          description: Required when dispatchMode is DELIVERY
        notes:
          type: string
          nullable: true
...
    DemandListing:
      type: object
      required:
        [id, vendorId, fishSpecies, marketLocation, quantityKg, offerPricePerKg, status, postedAt]
      properties:
        id:
          type: integer
          format: int64
        vendorId:
          type: integer
          format: int64
        vendorName:
          type: string
          nullable: true
        fishSpecies:
          $ref: '#/components/schemas/FishSpecies'
        marketLocation:
          $ref: '#/components/schemas/MarketLocation'
        quantityKg:
          type: number
          format: double
        offerPricePerKg:
          type: number
          format: double
        notes:
          type: string
          nullable: true
        neededBy:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/DemandListingStatus'
        postedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        distanceKm:
          type: number
          format: double
          nullable: true
          description: Haversine distance from buyer's lat/lng (when supplied) to listing market location
...
    BuyerVendorProfile:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        avgRating:
          type: number
          format: double
          nullable: true
          description: Average review rating (null until reviews are implemented)
        reviewCount:
          type: integer
          format: int64
          nullable: true
        joinedDate:
          type: string
          format: date-time
          nullable: true
        totalCompletedTrades:
          type: integer
          format: int64
          nullable: true
...
    FavoriteTargetType:
      type: string
      enum: [LISTING, VENDOR]
    BuyerFavorite:
      type: object
      required: [id, targetType, targetId, createdAt]
      properties:
        id:
          type: integer
          format: int64
        targetType:
          $ref: '#/components/schemas/FavoriteTargetType'
        targetId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        listing:
          allOf:
            - $ref: '#/components/schemas/DemandListing'
          nullable: true
          description: Populated when targetType=LISTING and the listing still exists
        vendor:
          allOf:
            - $ref: '#/components/schemas/BuyerVendorProfile'
          nullable: true
          description: Populated when targetType=VENDOR and the vendor still exists
    AddFavoriteRequest:
      type: object
      required: [targetType, targetId]
      properties:
        targetType:
          $ref: '#/components/schemas/FavoriteTargetType'
        targetId:
          type: integer
          format: int64
...
    UserRef:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
...
    Order:
      type: object
      required: [id, buyer, seller, species, agreedPricePerKg, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        buyer:
          $ref: '#/components/schemas/UserRef'
        seller:
          $ref: '#/components/schemas/UserRef'
        species:
          $ref: '#/components/schemas/FishSpecies'
        catchAlertId:
          type: integer
          format: int64
          nullable: true
        demandListingId:
          type: integer
          format: int64
          nullable: true
        agreedPricePerKg:
          type: number
          format: double
        orderedQtyEstimate:
          type: string
          nullable: true
        orderedQtyKg:
          type: number
          format: double
          nullable: true
        dispatchMode:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED, DISPUTED]
        notes:
          type: string
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
        handoff:
          $ref: '#/components/schemas/HandoffConfirmation'
          nullable: true
        payment:
          $ref: '#/components/schemas/PaymentRecord'
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    HandoffConfirmation:
      type: object
      required: [id, orderId, actualQtyKg, finalPricePerKg, totalAmount, confirmedByBuyer, confirmedBySeller, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        actualQtyKg:
          type: number
          format: double
        finalPricePerKg:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        confirmedByBuyer:
          type: boolean
        confirmedBySeller:
          type: boolean
        disputeReason:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, CONFIRMED, DISPUTED]
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    PaymentRecord:
      type: object
      required: [id, orderId, amount, method, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        handoffId:
          type: integer
          format: int64
          nullable: true
        amount:
          type: number
          format: double
        method:
          type: string
        proofReference:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, CONFIRMED, DISPUTED]
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        code:
          type: string
          nullable: true
        fieldErrors:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    OrderStatusEvent:
      type: object
      required: [id, orderId, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        status:
          type: string
          description: The status the order transitioned to
        actorId:
          type: integer
          format: int64
          nullable: true
        actorName:
          type: string
          nullable: true
          description: Resolved name of the actor who triggered the transition
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    UploadResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Public URL of the uploaded file
    Notification:
      type: object
      required: [id, type, title, body, createdAt]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [ORDER_STATUS, NEW_LISTING, MESSAGE, REVIEW_REMINDER, ADVISORY]
        title:
          type: string
        body:
          type: string
        link:
          type: string
          nullable: true
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    BuyerProfile:
      type: object
      required: [id, fullName, email, role]
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        avatarUrl:
          type: string
          nullable: true
        totalOrders:
          type: integer
        totalReviews:
          type: integer
        totalFavorites:
          type: integer
        memberSince:
          type: string
          format: date-time
    BuyerProfileUpdateRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 200
        avatarUrl:
          type: string
          nullable: true
    PaymentIntentResponse:
      type: object
      required: [clientKey, publicKey, gateway]
      properties:
        clientKey:
          type: string
        publicKey:
          type: string
        gateway:
          type: string
          enum: [PAYMONGO, STUB]